backend/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# RAG & LLM
langchain==0.2.6
langchain-community==0.2.6
langchain-ollama==0.1.1
langchain-huggingface==0.0.3

# Vector Store
chromadb==0.5.3

# Embeddings
sentence-transformers==3.0.1

# Document Parsing
pymupdf==1.24.5
python-docx==1.1.2
markdown==3.6

# Utilities
python-dotenv==1.0.1
pydantic==2.7.4
tqdm==4.66.4
numpy==1.26.4

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
